# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Renhead API",
        "description": "Class Controller",
        "version": "1.0.0"
    },
    "paths": {
...
        "/travel-payments": {
            "get": {
                "tags": [
                    "Travel Payments"
                ],
                "summary": "Get",
                "operationId": "f5d37d62809198e0a698cc2eba43a867",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Travel Payments"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Travel Payments"
                ],
                "operationId": "UpdateTravelPayment",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "description": "Amount",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Travel Payments"
                ],
                "operationId": "StoreTravelPayment",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "description": "Amount",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/travel-payments/{id}": {
            "get": {
                "tags": [
                    "Travel Payments"
                ],
                "summary": "Get",
                "operationId": "716359b98c9f8cd7e8aebb7424afc3f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Travel Payments"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Travel Payments"
                ],
                "summary": "Delete",
                "operationId": "DeleteTravelPayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Travel Payment ID to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}